elementView refactor
